Crate spl_discriminator

Crate defining a discriminator type, which creates a set of bytes meant to be unique for instructions or struct types

Re-exports

pub use discriminator::ArrayDiscriminator;

pub use discriminator::SplDiscriminate;

Modules

discriminator

Exports the discriminator module The traits and types used to create a discriminator for a type

Derive Macros

SplDiscriminate

Derive macro library to implement the SplDiscriminate trait on an enum or struct